size type changed back to int from long in "MaxSizeConfig"

+++ hazelcast-hibernate/hazelcast-hibernate3/src/main/java/com/hazelcast/hibernate/local/LocalRegionCache.java
@@ -232,8 +232,7 @@ public class LocalRegionCache implements RegionCache {
         final int maxSize;
         final long timeToLive;
         if (config != null) {
-            // For max-size config on map, we assume and check that size can be Integer.MAX_VALUE at most
-            maxSize = (int) config.getMaxSizeConfig().getSize();
+            maxSize = config.getMaxSizeConfig().getSize();
             timeToLive = config.getTimeToLiveSeconds() * SEC_TO_MS;
         } else {
             maxSize = MAX_SIZE;
